requirements.txt
fastapi>=0.115.12
uvicorn[standard]>=0.25.0
dagger-io==0.18.5
pydantic>=2.5.0
python-dotenv>=1.0.0
anyio>=4.2.0
numpy>=1.24.0
openai>=1.0.0
qdrant-client>=1.7.0
requests>=2.31.0
sentence-transformers>=2.3.0 